Move newly added helper functions to `NtUnicodeStr`
