+++ src/helpers.rs
@@ -2,11 +2,6 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 use core::cmp::Ordering;
-use core::mem;
-
-use widestring::U16CStr;
-
-use crate::error::{NtStringError, Result};
 
 /// Generic memory layout unified for `ANSI_STRING`, `OEM_STRING`, `UNICODE_STRING`,
 /// in their mutable and immutable versions.
@@ -19,58 +14,6 @@ pub(crate) struct RawNtString<T> {
     pub(crate) maximum_length: u16,
     /// String buffer
     pub(crate) buffer: T,
-}
-
-pub(crate) fn check_from_u16(buffer: &[u16]) -> Result<u16> {
-    let elements = buffer.len();
-    let length_usize = elements
-        .checked_mul(mem::size_of::<u16>())
-        .ok_or(NtStringError::BufferSizeExceedsU16)?;
-    let length = u16::try_from(length_usize).map_err(|_| NtStringError::BufferSizeExceedsU16)?;
-    Ok(length)
-}
-
-pub(crate) fn check_from_u16_until_nul(buffer: &[u16]) -> Result<(u16, u16)> {
-    let length;
-    let maximum_length;
-
-    match buffer.iter().position(|x| *x == 0) {
-        Some(nul_pos) => {
-            // Include the terminating NUL character in `maximum_length` ...
-            let maximum_elements = nul_pos
-                .checked_add(1)
-                .ok_or(NtStringError::BufferSizeExceedsU16)?;
-            let maximum_length_usize = maximum_elements
-                .checked_mul(mem::size_of::<u16>())
-                .ok_or(NtStringError::BufferSizeExceedsU16)?;
-            maximum_length = u16::try_from(maximum_length_usize)
-                .map_err(|_| NtStringError::BufferSizeExceedsU16)?;
-
-            // ... but not in `length`
-            length = maximum_length - mem::size_of::<u16>() as u16;
-        }
-        None => return Err(NtStringError::NulNotFound),
-    };
-
-    Ok((length, maximum_length))
-}
-
-pub(crate) fn check_from_u16_cstr(u16cstr: &U16CStr) -> Result<(u16, u16)> {
-    let buffer = u16cstr.as_slice_with_nul();
-
-    // Include the terminating NUL character in `maximum_length` ...
-    let maximum_length_in_elements = buffer.len();
-    let maximum_length_in_bytes = maximum_length_in_elements
-        .checked_mul(mem::size_of::<u16>())
-        .ok_or(NtStringError::BufferSizeExceedsU16)?;
-    let maximum_length =
-        u16::try_from(maximum_length_in_bytes).map_err(|_| NtStringError::BufferSizeExceedsU16)?;
-
-    // ... but not in `length`
-    debug_assert!(maximum_length >= mem::size_of::<u16>() as u16);
-    let length = maximum_length - mem::size_of::<u16>() as u16;
-
-    Ok((length, maximum_length))
 }
 
 /// Compare any two `u16` iterators and return an [`Ordering`] value.
