+++ src/unicode_string/strmut.rs
@@ -1,4 +1,4 @@
-// Copyright 2023 Colin Finck <colin@reactos.org>
+// Copyright 2023-2026 Colin Finck <colin@reactos.org>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 use core::cmp::Ordering;
@@ -9,7 +9,7 @@ use core::{fmt, mem, slice};
 use widestring::{U16CStr, U16Str};
 
 use crate::error::Result;
-use crate::helpers::{check_from_u16, check_from_u16_cstr, check_from_u16_until_nul, RawNtString};
+use crate::helpers::RawNtString;
 use crate::NtStringError;
 
 use super::{impl_eq, impl_partial_cmp, NtUnicodeStr};
@@ -122,7 +122,7 @@ impl<'a> NtUnicodeStrMut<'a> {
     ///
     /// [`try_from_u16_until_nul`]: Self::try_from_u16_until_nul
     pub fn try_from_u16(buffer: &mut [u16]) -> Result<Self> {
-        let length = check_from_u16(buffer)?;
+        let length = NtUnicodeStr::try_length_from_u16(buffer)?;
 
         Ok(Self {
             raw: RawNtString {
@@ -150,7 +150,7 @@ impl<'a> NtUnicodeStrMut<'a> {
     ///
     /// [`try_from_u16`]: Self::try_from_u16
     pub fn try_from_u16_until_nul(buffer: &mut [u16]) -> Result<Self> {
-        let (length, maximum_length) = check_from_u16_until_nul(buffer)?;
+        let (length, maximum_length) = NtUnicodeStr::try_length_from_u16_until_nul(buffer)?;
 
         Ok(Self {
             raw: RawNtString {
@@ -196,7 +196,7 @@ impl<'a> TryFrom<&'a mut U16CStr> for NtUnicodeStrMut<'a> {
     /// The internal buffer will be NUL-terminated.
     /// See the [module-level documentation](super) for the implications of that.
     fn try_from(value: &'a mut U16CStr) -> Result<Self> {
-        let (length, maximum_length) = check_from_u16_cstr(value)?;
+        let (length, maximum_length) = NtUnicodeStr::try_length_from_u16_cstr(value)?;
 
         Ok(Self {
             raw: RawNtString {
